backend/core/broker.py
"""
Broker Connection — Angel One SmartAPI Wrapper
===============================================
Handles authentication, order placement, and position queries.

TODO (Phase 2): Implement with real SmartAPI connection.
Currently returns mock data for development.
"""

import logging
from typing import Optional

logger = logging.getLogger(__name__)


class BrokerConnection:
    """Wrapper around Angel One SmartAPI."""

    def __init__(self, broker_config):
        self.config = broker_config
        self.session = None
        self.is_connected = False

    def connect(self) -> bool:
        """
        Authenticate with Angel One.
        
        TODO: Replace with real implementation:
            from SmartApi import SmartConnect
            import pyotp
            
            self.session = SmartConnect(api_key=self.config.api_key)
            totp = pyotp.TOTP(self.config.totp_secret).now()
            data = self.session.generateSession(
                self.config.client_id,
                self.config.password,
                totp
            )
        """
        logger.info("🔌 [MOCK] Broker connection established")
        self.is_connected = True
        return True

    def disconnect(self):
        """Close broker connection."""
        self.is_connected = False
        logger.info("🔌 Broker disconnected")

    def place_order(self, stock: str, token: str, direction: str,
                    quantity: int, price: float, order_type: str = "LIMIT") -> Optional[str]:
        """
        Place an order with the broker.
        Returns order_id if successful, None if failed.
        
        TODO: Replace with real SmartAPI order placement.
        """
        logger.info(f"📝 [MOCK] Order placed: {direction} {quantity}x {stock} @ ₹{price:.2f}")
        return f"MOCK_ORDER_{stock}_{direction}"

    def cancel_order(self, order_id: str) -> bool:
        """Cancel a pending order."""
        logger.info(f"❌ [MOCK] Order cancelled: {order_id}")
        return True

    def get_positions(self) -> list:
        """Get current open positions."""
        return []

    def get_ltp(self, token: str) -> float:
        """Get last traded price for a stock."""
        return 0.0

    def get_order_status(self, order_id: str) -> str:
        """Check order status: PENDING, EXECUTED, CANCELLED, REJECTED."""
        return "EXECUTED"
