backend/requirements.txt
# Core
pandas>=2.0.0
numpy>=1.24.0

# Angel One Broker API
smartapi-python>=1.3.0

# Technical Indicators
pandas-ta>=0.3.14

# Real-time Data
websocket-client>=1.6.0

# Firebase (data bridge to dashboard)
firebase-admin>=6.2.0

# Configuration
python-dotenv>=1.0.0

# Scheduling
schedule>=1.2.0

# TOTP for Angel One login
pyotp>=2.9.0

# Utilities
requests>=2.31.0
tabulate>=0.9.0
